requirements.txt
certifi==2020.4.5.1
dj-database-url==0.5.0
Django==2.2.5
django-cors-headers==3.5.0
djangorestframework==3.10.3
djangorestframework-api-key==2.0.0
entrypoints==0.3
gast==0.3.3
grpcio==1.29.0
gunicorn==19.9.0
h5py==2.10.0
optional-django==0.1.0
Pillow==7.1.2
pipenv==2020.6.2
pluggy==0.13.1
protobuf==3.12.0
psycopg2-binary==2.7.7
py==1.9.0
pyzmq==19.0.1
requests==2.23.0
requests-oauthlib==1.3.0
sqlparse==0.3.0
virtualenv==20.0.21
virtualenv-clone==0.5.4
webencodings==0.5.1
websocket-client==0.48.0
